refactor: auth·agent 도메인 패키지 구조 적용
