+++ app/domains/auth/controller/router.py
@@ -1,8 +1,8 @@
 from fastapi import APIRouter, Cookie, Depends, HTTPException, Request, Response, status
 
-from app.domains.auth.dependencies import get_auth_service, get_current_user
+from app.domains.auth.dependency.dependencies import get_auth_service, get_current_user
 from app.domains.auth.models import User
-from app.domains.auth.schemas import (
+from app.domains.auth.schema.schemas import (
     AuthTokenResponse,
     LoginRequest,
     MessageResponse,
@@ -10,7 +10,7 @@ from app.domains.auth.schemas import (
     SignupRequest,
     UserResponse,
 )
-from app.domains.auth.service import (
+from app.domains.auth.service.service import (
     AuthError,
     AuthService,
     DuplicateUserError,
