+++ app/domains/classroom/controller/student_class_router.py
@@ -9,7 +9,7 @@
 """
 from fastapi import APIRouter, Depends
 
-from app.domains.auth.dependencies import get_current_user
+from app.domains.auth.dependency.dependencies import get_current_user
 from app.domains.auth.models import User
 from app.domains.classroom.dependency.dependencies import get_classroom_service
 from app.domains.classroom.schema.schemas import StudentMyClassResponse
