+++ app/domains/progress/controller/router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends
 
-from app.domains.auth.dependencies import get_current_student
-from app.domains.auth.dependencies import get_current_user
+from app.domains.auth.dependency.dependencies import get_current_student
+from app.domains.auth.dependency.dependencies import get_current_user
 from app.domains.auth.models import User
 from app.domains.content.dependency.dependencies import get_content_catalog_service
 from app.domains.content.service.service import ContentCatalogService
