+++ app/domains/progress/controller/teacher_router.py
@@ -9,7 +9,7 @@ progress 도메인에 위치한 이유:
 """
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 
-from app.domains.auth.dependencies import get_current_teacher
+from app.domains.auth.dependency.dependencies import get_current_teacher
 from app.domains.auth.models import User
 from app.domains.classroom.dependency.dependencies import get_classroom_service
 from app.domains.classroom.service.service import ClassroomService
