+++ tests/structure/test_domain_structure.py
@@ -10,10 +10,10 @@ def test_domain_packages_are_importable():
         "app.domains.developer.router",
         "app.domains.auth",
         "app.domains.auth.models",
-        "app.domains.auth.repository",
-        "app.domains.auth.router",
-        "app.domains.auth.schemas",
-        "app.domains.auth.service",
+        "app.domains.auth.repository.repository",
+        "app.domains.auth.controller.router",
+        "app.domains.auth.schema.schemas",
+        "app.domains.auth.service.service",
         "app.domains.classroom.controller.teacher_class_router",
         "app.domains.classroom.controller.student_class_router",
         "app.domains.progress.controller.teacher_router",
@@ -22,8 +22,8 @@ def test_domain_packages_are_importable():
         "app.domains.classroom.repository.repository",
         "app.domains.classroom.service.service",
         "app.domains.classroom.schema.schemas",
-        "app.domains.agent.schemas",
-        "app.domains.agent.service",
+        "app.domains.agent.schema.schemas",
+        "app.domains.agent.service.service",
         "app.domains.instruction.controller.student_router",
         "app.domains.instruction.controller.teacher_router",
         "app.domains.instruction.service.generation",
@@ -53,8 +53,8 @@ def test_domain_packages_are_importable():
         "app.domains.content.stage3.schemas",
         "app.domains.content.stage3.service",
         "app.domains.agent",
-        "app.domains.agent.router",
-        "app.domains.agent.schemas",
+        "app.domains.agent.controller.router",
+        "app.domains.agent.schema.schemas",
         "app.infrastructure.rag",
         "app.infrastructure.rag.retriever",
         "app.infrastructure.rag.schemas",
